Fix the definition of "struct stat" on 32-bit FreeBSD 12+
The original definitions were never correct.  But nobody noticed because
we don't do CI on 32-bit FreeBSD.  The problem is apparent now due to
 #3723 , which caused the nightly toolchain to switch to a FreeBSD 12
ABI.

Fixes https://github.com/rust-lang/rust/issues/130677
