+++ src/unix/bsd/freebsdlike/freebsd/freebsd11/mod.rs
@@ -479,10 +479,11 @@ extern "C" {
 }
 
 cfg_if! {
-    if #[cfg(any(target_arch = "x86_64",
-                 target_arch = "aarch64",
-                 target_arch = "riscv64"))] {
+    if #[cfg(target_pointer_width = "64")] {
         mod b64;
         pub use self::b64::*;
+    } else {
+        mod b32;
+        pub use self::b32::*;
     }
 }
